backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.6.0
python-dotenv==1.0.0
openai==1.6.0
PyPDF2==3.0.1
python-docx==0.8.11
aiofiles==23.2.1
